% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YuriR.R
\docType{data}
\name{cols}
\alias{cols}
\title{Yuri}
\format{
An object of class \code{character} of length 10.
}
\usage{
cols
}
\arguments{
\item{cols}{}

\item{fills}{}

\item{t_theme}{}

\item{outer.IQR}{}
}
\description{
主题+一些函数
}
\examples{

}
\keyword{datasets}
